+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+- Make attachment updates work on both Confluence Server and Confluence Cloud
+
 ## 1.0.4 - 2021-11-25
 - Remove mentions of non-existing force-unique parameter
 - Bump dependencies